Move middlewares to root dir.
