+++ lib/thin.rb
@@ -1,3 +1,23 @@
-require "thin/version"
-require "thin/system"
-require "thin/server"
+module Thin
+  module Backends
+    autoload :Prefork, "thin/backends/prefork"
+    autoload :SingleProcess, "thin/backends/single_process"
+  end
+
+  autoload :Async, "thin/async"
+  autoload :CatchAsync, "thin/catch_async"
+  autoload :Chunked, "thin/chunked"
+  autoload :Configurator, "thin/configurator"
+  autoload :Connection, "thin/connection"
+  autoload :FastEnumerator, "thin/fast_enumerator"
+  autoload :Listener, "thin/listener"
+  autoload :Request, "thin/request"
+  autoload :Response, "thin/response"
+  autoload :Runner, "thin/runner"
+  autoload :Server, "thin/server"
+  autoload :StreamFile, "thin/stream_file"
+  autoload :Streamed, "thin/streamed"
+  autoload :System, "thin/system"
+  autoload :Threaded, "thin/threaded"
+  autoload :Version, "thin/version"
+end
