+++ lib/thin/catch_async.rb
