+++ lib/thin/threaded.rb
@@ -1,0 +1,17 @@
+module Thin
+  # Has to be the first middleware
+  class Threaded
+    def initialize(app, options={})
+      @app = app
+      EM.threadpool_size = options[:pool_size] if options.key?(:pool_size)
+    end
+
+    def call(env)
+      env['rack.multithread'] = true
+
+      EM.defer(proc { @app.call(env) }, env['thin.process'])
+
+      [-1, {}, []]
+    end
+  end
+end
